pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opsdroid-skill-networkinfo"
version = "0.0.24"
authors = [
    { name="Darren Spruell", email="phatbuckett@gmail.com" },
]
description = "Network information skill for Opsdroid"
readme = "README.md"
dependencies = [
    "aslookup==1.5.0",
    "defang==0.5.3",
    "dnspython==2.3.0",
    "ipapi==1.0.4",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 2 - Pre-Alpha",
    # "Development Status :: 3 - Alpha",
    # "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    # "Development Status :: 6 - Mature",
    # "Development Status :: 7 - Inactive",
]

[project.urls]
"Homepage" = "https://github.com/dspruell/skill-networkinfo"

[project.entry-points."opsdroid_skills"]
networkinfo = "networkinfo"
